changes after cr
