+++ packages/mobile/src/features/videos/components/Movies/SlaiderLarge.tsx
@@ -9,12 +9,17 @@ import {
 import { GetVideosList } from '../../action/conector';
 import { NavigationName } from '../../../../config/routing/NavigationName';
 import { TouchableOpacity } from 'react-native-gesture-handler';
+import { AnimationSlaider } from './AnimationSlaider';
 
  type SlaiderLargeS = {
   listVideos: Array<MoviesListSimpleType>
  }
 
-export class SlaiderLarge extends React.Component<any,SlaiderLargeS>{
+ type SlaiderLargeP = {
+   navigation: any
+ }
+
+export class SlaiderLarge extends React.Component<SlaiderLargeP,SlaiderLargeS>{
   x: Animated.Value;
   onScroll: ()=>void;
   renderIt: (item,index)=>void;
@@ -24,7 +29,7 @@ export class SlaiderLarge extends React.Component<any,SlaiderLargeS>{
       listVideos : []
     }
     this.x = new Animated.Value(10);
-    this.renderIt = ({item, index}) => this.ImgeSlaider(item, index, this.x);
+    this.renderIt = ({item, index}) => <AnimationSlaider item={item} index={index} x={this.x} navigation={this.props.navigation} />;
     this.onScroll = Animated.event(
       [{nativeEvent: { contentOffset: { x: this.x} } } ], 
       { useNativeDriver: true}
@@ -38,35 +43,7 @@ export class SlaiderLarge extends React.Component<any,SlaiderLargeS>{
     })
   }
 
-  public navToSingleMovie = () => {this.props.navigation.navigate(NavigationName.SINGLEMOVIE)}
-
   public Separator = () => (<ViewSeparator/>)
-  public ImgeSlaider = (item:MoviesListSimpleType, index, x:Animated.Value) => {
-    const { width } = Dimensions.get('window')
-    const wCard = width -90 ;
-    const position = Animated.subtract(index * wCard,x);
-    const isDisappearing = -width;
-    const isLeft = 0;
-    const isRight =  width - 20;
-    const isAppearing = width * 2;
-    const scale = position.interpolate({
-      inputRange: [isDisappearing, isLeft, isRight, isAppearing],
-      outputRange: [0.4,1,0.8,0.8],
-      extrapolate: 'clamp'
-    })
-    const opacity = position.interpolate({
-      inputRange: [isDisappearing, isLeft, isRight, isAppearing],
-      outputRange: [0.4,1,0.3,0.5],
-    })
-
-    return(
-      <TouchableOpacity onPress={this.navToSingleMovie}>
-        <Animated.View >
-          <ImageSlaider style={{ opacity, transform: [{ scale }]}} source={{uri: item.uri}}/>
-        </Animated.View>
-      </TouchableOpacity>
-    )
-  }
 
     render(){
         return(
